% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.r
\name{ConnectionPool}
\alias{ConnectionPool}
\title{Connection Pool implementation}
\description{
the classical and notorious connecton pool.
It needs a \code{facotry} (a.k.a. a function) which creates
the actual DB connections
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ConnectionPool-new}{\code{ConnectionPool$new()}}
\item \href{#method-ConnectionPool-get_connection}{\code{ConnectionPool$get_connection()}}
\item \href{#method-ConnectionPool-release}{\code{ConnectionPool$release()}}
\item \href{#method-ConnectionPool-clone}{\code{ConnectionPool$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConnectionPool-new"></a>}}
\if{latex}{\out{\hypertarget{method-ConnectionPool-new}{}}}
\subsection{Method \code{new()}}{
builds a Connection Pool
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionPool$new(factory, min = 1, max = 5)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{factory}}{actual function (callable) which creates the connection}

\item{\code{min}}{minimun number of connections (default: 1)}

\item{\code{max}}{maximum number of connections (default: 5)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConnectionPool-get_connection"></a>}}
\if{latex}{\out{\hypertarget{method-ConnectionPool-get_connection}{}}}
\subsection{Method \code{get_connection()}}{
returns a PooledConnection
the only thing a PooledConnection does is to return to the pool
when \code{DBI::dbDisconnect} is called on it
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionPool$get_connection()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConnectionPool-release"></a>}}
\if{latex}{\out{\hypertarget{method-ConnectionPool-release}{}}}
\subsection{Method \code{release()}}{
releases a Connection to the pool
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionPool$release(con)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{con}}{to be released}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConnectionPool-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ConnectionPool-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectionPool$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
