% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.r
\name{dbExistsTable,PooledConnection,character-method}
\alias{dbExistsTable,PooledConnection,character-method}
\title{Delegate for dbExistsTable}
\description{
Calls the underlaying methtod on the DBIConnection object
}
