% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.r
\name{dbGetQuery,PooledConnection,character-method}
\alias{dbGetQuery,PooledConnection,character-method}
\title{Delegate for dbGetQuery}
\description{
Calls the underlaying methtod on the DBIConnection object
}
