% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.r
\name{dbListFields,PooledConnection,character-method}
\alias{dbListFields,PooledConnection,character-method}
\title{Delegate for dbListFields}
\description{
Calls the underlaying methtod on the DBIConnection object
}
