% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.r
\name{dbQuoteIdentifier,PooledConnection,ANY-method}
\alias{dbQuoteIdentifier,PooledConnection,ANY-method}
\title{Delegate for dbQuoteIdentifier}
\description{
Calls the underlaying methtod on the DBIConnection object
}
