% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.r
\name{dbRemoveTable,PooledConnection,character-method}
\alias{dbRemoveTable,PooledConnection,character-method}
\title{Delegate for dbRemoveTable}
\description{
Calls the underlaying methtod on the DBIConnection object
}
