% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.r
\name{dbSendQuery,PooledConnection-method}
\alias{dbSendQuery,PooledConnection-method}
\title{Delegate for dbSendQuery}
\arguments{
\item{conn}{conn PooledConnection object with DBIConnection slot}

\item{...}{generic arguments passed to DBIConnectiton}
}
\description{
Calls the underlaying methtod on the DBIConnection object
}
\seealso{
DBI::dbSendQuery
}
