% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.r
\name{dbUnquoteIdentifier,PooledConnection,SQL-method}
\alias{dbUnquoteIdentifier,PooledConnection,SQL-method}
\title{Delegate for dbUnquoteIdentifier}
\description{
Calls the underlaying methtod on the DBIConnection object
}
